submodules/modelAPI/requirements.txt
numpy==1.16.2
opencv-python==4.0.0.21
keras
Pillow
matplotlib